Extract FullscreenViewer and ToggleFullscreenCornerButton
Allow displaying other elements besides images in overlays that can be
closed via back button or escape key.

Prevent duplicating styles to display a toggle fullscreen button with
a gradient in an element's corner.

REDMINE-20325
