+++ entry_types/scrolled/package/src/frontend/ExpandableImage/Viewer.js
@@ -1,85 +1,31 @@
-import React, {useCallback, useState, useEffect} from 'react'
-import classNames from 'classnames';
-import {useDelayedBoolean} from '../useDelayedBoolean';
-import {Fullscreen} from './Fullscreen';
+import React from 'react'
 import {ZoomableImage} from './ZoomableImage';
-import {ToggleFullscreenButton} from '../ToggleFullscreenButton';
+import {ToggleFullscreenCornerButton} from '../ToggleFullscreenCornerButton';
 import {useContentElementEditorState} from '../useContentElementEditorState';
 
-import styles from './Viewer.module.css';
+import {FullscreenViewer} from '../FullscreenViewer';
 
 export default function Viewer({
   imageFile,
   contentElementId,
   children
 }) {
-  const [isFullscreen, setIsFullscreen] = useState(false);
-  const isRendered = useDelayedBoolean(isFullscreen, {fromTrueToFalse: 200});
-  const isVisible = useDelayedBoolean(isFullscreen, {fromFalseToTrue: 1});
-
   const {isEditable, isSelected} = useContentElementEditorState();
 
-  useEffect(() => {
-    function handlePopState() {
-      setIsFullscreen(
-        window.history.state?.fullscreenInlineImage === contentElementId
-      );
-    }
-
-    window.addEventListener('popstate', handlePopState);
-    return () => window.removeEventListener('popstate', handlePopState);
-  });
-
-  const enterFullscreen = useCallback(() => {
-    setIsFullscreen(true);
-
-    if (window.parent === window) {
-      window.history.pushState({fullscreenInlineImage: contentElementId}, '');
-    }
-  }, [contentElementId]);
-
-  const exitFullscreen = useCallback(() => {
-    setIsFullscreen(false);
-
-    if (window.history.state?.fullscreenInlineImage && window.parent === window) {
-      window.history.back();
-    }
-  }, []);
-
-  useEffect(() => {
-    function handleKeyDown(event) {
-      if (event.key === 'Escape') {
-        exitFullscreen();
-      }
-    }
-
-    window.addEventListener('keydown', handleKeyDown);
-    return () => window.removeEventListener('keydown', handleKeyDown);
-  }, [exitFullscreen]);
-
   return (
-    <>
-      <div onClick={enterFullscreen}
-           style={{pointerEvents: isEditable && !isSelected ? 'none' : undefined}}>
-        {children}
-        <div className={styles.controls}>
-          <ToggleFullscreenButton isFullscreen={false}
-                                  onEnter={enterFullscreen}
-                                  onExit={exitFullscreen} />
+    <FullscreenViewer
+      contentElementId={contentElementId}
+      renderChildren={({enterFullscreen}) =>
+        <div onClick={enterFullscreen}
+             style={{pointerEvents: isEditable && !isSelected ? 'none' : undefined}}>
+          {children}
+          <ToggleFullscreenCornerButton isFullscreen={false}
+                                        onEnter={enterFullscreen} />
         </div>
-      </div>
-      {isRendered &&
-       <Fullscreen>
-         <div className={classNames(styles.wrapper, {[styles.visible]: isVisible})}>
-           <ZoomableImage onClose={exitFullscreen}
-                          imageFile={imageFile} />
-           <div className={styles.controls}>
-             <ToggleFullscreenButton isFullscreen={true}
-                                     onEnter={enterFullscreen}
-                                     onExit={exitFullscreen} />
-           </div>
-         </div>
-       </Fullscreen>}
-    </>
+      }
+      renderFullscreenChildren={({exitFullscreen}) =>
+        <ZoomableImage onClose={exitFullscreen}
+                       imageFile={imageFile} />
+      } />
   );
 }
