+++ entry_types/scrolled/package/src/frontend/FullscreenViewer/Fullscreen.js
