+++ entry_types/scrolled/package/src/frontend/Panorama/Viewer.module.css
@@ -9,20 +9,6 @@
   cursor: move;
 }
 
-.controls {
-  position: absolute;
-  right: 0;
-  top: 0;
-  color: #fff;
-  background: linear-gradient(30deg, transparent, transparent 60%, #000000a6);
-  padding: 5px 5px 40px 70px;
-  pointer-events: none;
-}
-
-.controls > * {
-  pointer-events: auto;
-}
-
 .spinner {
   animation: spin 0.75s linear infinite;
   width: 40px;
