+++ entry_types/scrolled/package/src/frontend/ToggleFullscreenCornerButton.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import {ToggleFullscreenButton} from './ToggleFullscreenButton';
+
+import styles from './ToggleFullscreenCornerButton.module.css';
+
+export function ToggleFullscreenCornerButton(props) {
+  return (
+    <div className={styles.corner}>
+      <ToggleFullscreenButton {...props} />
+    </div>
+  );
+}
