+++ entry_types/scrolled/package/src/frontend/ToggleFullscreenCornerButton.module.css
@@ -1,0 +1,13 @@
+.corner {
+  position: absolute;
+  right: 0;
+  top: 0;
+  color: #fff;
+  background: linear-gradient(30deg, transparent, transparent 60%, #000000a6);
+  padding: 5px 5px 40px 70px;
+  pointer-events: none;
+}
+
+.corner > * {
+  pointer-events: auto;
+}
